Replaced void* with boost::any
Makes default/implicitly created copy constructors well behaved
Implicit destructor now possible
which makes move constructor available and implicitly created
by using the any we are compliant with the rule of 0

Note on moved from state:
Move operations move the contents of the any, and not necessarily the container it contains.  So assignment operators and clear would return it to a specified state.  Which is to my understanding the requirement for "valid but unspecified".  All other operations will throw, until it is assigned to or clear() is run.

If more functionality is desired post move we would have to declare/define all the constructors and assignment operators.